+++ backend/package.json
@@ -1,0 +1,53 @@
+{
+  "name": "my-zone-backend",
+  "version": "1.0.0",
+  "description": "my zone web app backend server",
+  "main": "index.js",
+  "scripts": {
+    "start": "tsc && cross-env NODE_ENV=development node dist/main.js",
+    "start:pro": "tsc && cross-env NODE_ENV=production node dist/main.js",
+    "build": "tsc",
+    "test": "jest"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/m544498510/maysApp.git"
+  },
+  "keywords": [],
+  "author": "May",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/m544498510/genealogyWebApp/issues"
+  },
+  "homepage": "https://github.com/m544498510/genealogyWebApp#readme",
+  "dependencies": {
+    "apollo-server-koa": "^2.6.1",
+    "graphql": "^14.3.1",
+    "koa": "^2.7.0",
+    "koa-bodyparser": "^4.2.1",
+    "koa-json": "^2.0.2",
+    "koa-logger": "^3.2.0",
+    "koa-onerror": "^4.1.0",
+    "koa-router": "^7.4.0",
+    "koa-static": "^5.0.0",
+    "log4js": "^4.3.1",
+    "moment": "^2.24.0",
+    "mongoose": "^5.5.12"
+  },
+  "devDependencies": {
+    "@types/graphql": "^14.2.0",
+    "@types/jest": "^24.0.13",
+    "@types/koa": "^2.0.48",
+    "@types/koa-bodyparser": "^4.3.0",
+    "@types/koa-json": "^2.0.18",
+    "@types/koa-router": "^7.0.40",
+    "@types/koa-static": "^4.0.1",
+    "@types/mongodb": "^3.1.28",
+    "@types/mongoose": "^5.5.3",
+    "@types/node": "^12.0.4",
+    "cross-env": "^5.2.0",
+    "jest": "^24.8.0",
+    "ts-jest": "^24.0.2",
+    "typescript": "^3.5.1"
+  }
+}
