+++ backend/src/modules/module1.ts
@@ -1,0 +1,16 @@
+import {makeExecutableSchema, gql} from 'apollo-server-koa';
+export default makeExecutableSchema({
+  typeDefs: gql`
+    type User {
+      name: String
+    }
+    type Query {
+      User: User
+    }
+  `,
+  resolvers: {
+    Query: {
+      User: () => ({ name: 'Hello world!'})
+    },
+  }
+});
