--- server/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "genealog-backend",
-  "version": "1.0.0",
-  "description": "genealogy web app backend server",
-  "main": "index.js",
-  "scripts": {
-    "start": "tsc && cross-env NODE_ENV=development node dist/main.js",
-    "start:pro": "tsc && cross-env NODE_ENV=production node dist/main.js",
-    "build": "tsc",
-    "test": "jest"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/m544498510/genealogyWebApp.git"
-  },
-  "keywords": [],
-  "author": "May",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/m544498510/genealogyWebApp/issues"
-  },
-  "homepage": "https://github.com/m544498510/genealogyWebApp#readme",
-  "dependencies": {
-    "apollo-server-koa": "^1.3.2",
-    "graphql": "^0.13.1",
-    "koa": "^2.5.0",
-    "koa-bodyparser": "^4.2.0",
-    "koa-json": "^2.0.2",
-    "koa-logger": "^3.1.0",
-    "koa-onerror": "^4.0.0",
-    "koa-router": "^7.4.0",
-    "koa-static": "^4.0.2",
-    "log4js": "^2.5.3",
-    "mongoose": "^5.0.7"
-  },
-  "devDependencies": {
-    "@types/graphql": "^0.12.4",
-    "@types/jest": "^22.1.4",
-    "@types/koa": "^2.0.44",
-    "@types/koa-bodyparser": "^4.2.0",
-    "@types/koa-json": "^2.0.18",
-    "@types/koa-router": "^7.0.27",
-    "@types/koa-static": "^4.0.0",
-    "@types/mongodb": "^3.0.5",
-    "@types/mongoose": "^5.0.3",
-    "@types/node": "^9.4.6",
-    "cross-env": "^5.1.3",
-    "jest": "^22.4.2",
-    "ts-jest": "^22.4.1",
-    "typescript": "^2.7.2"
-  }
-}
